Combine boosting/bagging trees into one file.
Removes a lot of code duplication.
